refactor plot_fundus into plot_axon_map, plot_implant_on_axon_map
